app/obolapi: implement auth v2
This is the second iteration of an authentication scheme for partial exits Obol API integration.

The main breaking change is that the GET endpoint to retrieve full exits data has changed, and now is:

```
/exp/exit/{lock_hash}/{share_index}/{validator_pubkey}
```

---

## `POST`ing a partial exit

The data type that client sends to API for a partial exit changed, and now is:

```json
{
    "partial_exits": [
        {
            "public_key": "validator public key, 0x-prefixed hex encoded"
            "signed_exit_message": {
                "message": {
                    "epoch": "validator exit epoch",
                    "validator_index": "validator index"
                },
                "signature": "signature of message as defined by the spec"
            }
        }
    ],
        "share_idx": 42, // share index of the operator signing this object
        "signature": "base64-encoded secp256k1 signature of the root of partial_exits and share_index"
}
```

The client now signs `partial_exit` and `share_idx` tree roots in this order, and signs it with its Charon identity key,
identified by the `share_idx` value.

The server will verify this signature before proceeding with any computation.

## `GET`ing the full exit message

To retrieve a signature, the client signs the lock hash of the cluster lock that contains the specified validator, with
the specific share index --- cluster lock hash and share index are specified in the URL itself.

Once signed, the client will use this value as a Bearer token, base64-encoded.

The server will verify this signature before proceeding with any computation.

The return data type of this endpoint did not change.