app/obolapi: ignore empty partial signatures
API will return partial signatures iff there are at least threshold of them, which also means we could find `null`'s in the response.

Since the partial signatures array returned by the API maps 1:1 with the corresponding share index, we need to ignore empty ones to keep going with the aggregation process.

We still error if the signature is longer than 0 bytes but no more than 2.

Added a test case to cover the issue as well.