*: swap obol api library, beacon node client
With Charon v1.0 we released sign commands that use a derivative work coming from this repo: unify the codebases by importing and using Charon's libraries in here.

Refactor beacon node client utilization to be aware of the cluster lock's fork hash.

Remove `mockserver` command as it is not being actively used anymore.